improve FunctionExpression code, add some test
