+++ .eslintrc
@@ -9,6 +9,7 @@
   },
   "rules": {
     "no-use-before-define": [2, "nofunc"],
-    "no-param-reassign": 0
+    "no-param-reassign": 0,
+		"new-cap": 0
   }
 }
